Fix dynamic region creation function example
Version 1.0.0-incubating.M3 of Geode added restrictions
on region names, one of which prevents the creation of
regions whose names begin with a double underscore.
The example code detailing how to create a dynamic
region on the server no longer works, so this commit
removes an underscore from the name of the region
attributes metadata region so that it is now valid.